requirements.txt
redis
psycopg2-binary
Faker
fastapi
uvicorn